Simplify theme handling
Instead of relying on a hack to choose the theme in CSS, set it
explicitly from the backend in a variable that will be read by the
templates.

The templates then can set a special CSS class name in the body
element, and all the CSS will be correctly applied. Also, remove a
now-unnecessary hack to make the dark theme cover the whole page.